Clase separate pentru admin
