+++ src/main/java/org/sportstogo/backend/Controller/AdminController.java
@@ -1,0 +1,23 @@
+package org.sportstogo.backend.Controller;
+
+import org.sportstogo.backend.Models.User;
+import org.sportstogo.backend.Service.AdminService;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping(path = "admin")
+public class AdminController {
+
+    private final AdminService adminService;
+
+    public AdminController(AdminService adminService) {
+        this.adminService = adminService;
+    }
+
+    @GetMapping(path = "recent-users")
+    public List<User> getRecentUsers() {
+        return this.adminService.getUsersRegisteredLastWeek();
+    }
+}
