+++ src/main/java/org/sportstogo/backend/Controller/UserController.java
@@ -40,9 +40,4 @@ public class UserController {
         return ResponseEntity.ok()
                 .body("User updated successfully");
     }
-    @GetMapping(path = "recent")
-    public List<User> getRecentUsers() {
-        return this.userService.getUsersRegisteredLastWeek();
-    }
-
 }
