+++ src/main/java/org/sportstogo/backend/Service/AdminService.java
@@ -1,0 +1,22 @@
+package org.sportstogo.backend.Service;
+
+import lombok.AllArgsConstructor;
+import org.sportstogo.backend.Models.User;
+import org.sportstogo.backend.Repository.AdminRepository;
+import org.sportstogo.backend.Repository.UserRepository;
+import org.springframework.stereotype.Service;
+
+import java.time.LocalDate;
+import java.util.List;
+
+@Service
+@AllArgsConstructor
+public class AdminService {
+
+    private final AdminRepository adminRepository;
+
+    public List<User> getUsersRegisteredLastWeek() {
+        LocalDate oneWeekAgo = LocalDate.now().minusDays(7);
+        return adminRepository.findUsersRegisteredAfter(oneWeekAgo);
+    }
+}
